Refactor AuthSelector
