+++ playground/src/garbage/RequestorPage/RequestPanel/AuthSelector/BearerAuthorizationItem.tsx
@@ -1,0 +1,15 @@
+import type { BearerAuthorization } from "../../store/slices/settingsSlice";
+import { MaskedText } from "./MaskedText";
+
+export const BearerAuthorizationItem = (
+  props: Pick<BearerAuthorization, "id" | "name" | "token">,
+) => {
+  return (
+    <div className="grid grid-cols-[auto_1fr] gap-2 items-center">
+      Bearer
+      <div className="font-mono">
+        <MaskedText text={props.token} />
+      </div>
+    </div>
+  );
+};
