+++ playground/src/garbage/RequestorPage/Settings/Auths/AuthForm.tsx
@@ -1,10 +1,17 @@
 import type { Authorization } from "../../store/slices/settingsSlice";
 import { BearerForm } from "./BearerForm";
 
-export function AuthForm(props: { auth: Authorization, isNew?: boolean }) {
+export function AuthForm(props: { auth: Authorization; isNew?: boolean }) {
   const { auth, isNew } = props;
   if (auth.type === "bearer") {
-    return <BearerForm id={auth.id} token={auth.token} name={auth.name} isNew={isNew} />;
+    return (
+      <BearerForm
+        id={auth.id}
+        token={auth.token}
+        name={auth.name}
+        isNew={isNew}
+      />
+    );
   }
 
   return <div>Not supported ({auth.type})</div>;
